# packages/api-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/ab/peer/add/{guid}": {
      "post": {
        "description": "HTTP 200 with an EMPTY body. The client treats 204 and any non-empty body as a failure.",
        "operationId": "RustdeskAb_addPeer",
        "parameters": [
          {
            "name": "guid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbPeerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Empty body"
          },
          "401": {
            "description": "The client drops its token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RustdeskErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Read-only book or peer limit reached",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RustdeskErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RustdeskErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a peer",
        "tags": [
          "rustdesk"
        ]
      }
    },
...
    "/api/ab/tag/add/{guid}": {
      "post": {
        "description": "HTTP 200 with an EMPTY body. The client treats 204 and any non-empty body as a failure.",
        "operationId": "RustdeskAb_addTag",
        "parameters": [
          {
            "name": "guid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbTagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Empty body"
          },
          "401": {
            "description": "The client drops its token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RustdeskErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RustdeskErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a tag (idempotent)",
        "tags": [
          "rustdesk"
        ]
      }
    },
...
    "/api/admin/address-books": {
      "get": {
        "operationId": "AdminAddressBooks_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "`field:asc` or `field:desc`; allowed fields are listed per endpoint",
            "schema": {
              "example": "createdAt:desc",
              "type": "string"
            }
          },
          {
            "name": "kind",
            "required": false,
            "in": "query",
            "description": "Sort fields: name, createdAt, updatedAt.",
            "schema": {
              "$ref": "#/components/schemas/AddressBookKind"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Matches the book name.",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressBookPageDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "List address books (personal and shared)",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminAddressBooks_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressBookDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressBookDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Create a shared address book (owned by the caller)",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/address-books/{guid}": {
      "get": {
        "operationId": "AdminAddressBooks_get",
        "parameters": [
          {
            "name": "guid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressBookDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Get an address book",
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminAddressBooks_update",
        "parameters": [
          {
            "name": "guid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressBookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressBookDto"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Rename an address book or change its note",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "description": "Personal books are deleted with their user (409 here).",
        "operationId": "AdminAddressBooks_delete",
        "parameters": [
          {
            "name": "guid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Delete a shared address book",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/admin/address-books/{guid}/shares": {
      "get": {
        "operationId": "AdminAddressBooks_shares",
        "parameters": [
          {
            "name": "guid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShareDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Users a book is shared with",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "description": "Body: `[{userId, rule}]`; rule 1 read-only, 2 read/write, 3 full control.",
        "operationId": "AdminAddressBooks_replaceShares",
        "parameters": [
          {
            "name": "guid",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ShareInputDto"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ShareDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or session expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an administrator, or Origin not allowed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "rd_admin_session": []
          }
        ],
        "summary": "Replace the share list of a shared book",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "RustDesk API server",
    "description": "Self-hosted replacement for the RustDesk **API server** (port 21114). It does not replace `hbbs`/`hbbr`.\n\n**rustdesk** — endpoints the RustDesk client calls. Quirks of the client contract:\n- bodies are parsed as JSON whatever the Content-Type; an empty body is `{}`;\n- errors are `{\"error\": \"<string>\"}`;\n- address-book mutations and audit posts answer **200 with an empty body** (204 or a non-empty body is a failure for the client);\n- `/api/sysinfo` and `/api/sysinfo_ver` answer plain text;\n- audit, heartbeat and sysinfo are unauthenticated by protocol.\n\n**admin** — the admin panel API under `/api/admin`: session cookie, role ADMIN, an allowed `Origin` on state-changing requests,\nerrors `{\"error\": {\"code\", \"message\", \"details\"?}}`, ISO-8601 UTC timestamps, durations in whole seconds.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token from POST /api/login"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rd_admin_session"
      }
    },
    "schemas": {
      "RustdeskErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Wrong username or password"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AbPeerDto": {
        "type": "object",
        "properties": {
          "forceAlwaysRelay": {
            "type": "string",
            "description": "\"true\" or \"false\" — a string in the client"
          },
          "same_server": {
            "type": "boolean"
          },
          "id": {
            "type": "string",
            "description": "RustDesk ID of the peer."
          },
          "alias": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "note": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "hash": {
            "type": "string",
            "description": "Saved-password hash (personal books). A credential: stored encrypted."
          },
          "password": {
            "type": "string",
            "description": "Shared password (shared books). A credential: stored encrypted."
          },
          "rdpPort": {
            "type": "string"
          },
          "rdpUsername": {
            "type": "string"
          },
          "loginName": {
            "type": "string"
          },
          "device_group_name": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "AbTagDto": {
        "type": "object",
        "properties": {
          "color": {
            "type": "number",
            "description": "32-bit ARGB color.",
            "minimum": -2147483648,
            "maximum": 4294967295,
            "example": 4283215696
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "color",
          "name"
        ]
      },
...
      "AdminErrorDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/AdminErrorBodyDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AddressBookKind": {
        "type": "string",
        "enum": [
          "PERSONAL",
          "SHARED"
        ]
      },
...
      "AddressBookPageDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressBookDto"
            }
          },
          "total": {
            "type": "number",
            "example": 1
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 50
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateAddressBookDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateAddressBookDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "nullable": true,
            "maxLength": 2000
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          }
        }
      },
      "ShareDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "rule": {
            "enum": [
              1,
              2,
              3
            ],
            "type": "number",
            "description": "1 read-only, 2 read/write, 3 full control"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "rule",
          "username"
        ]
      },
...
    }
  }
}